fix a bug with the rendering of the offer
